Widen sectors_in_ssd to uint64_t to avoid overflow on large disks
sectors_in_ssd was a uint32_t, computed as
sectors_per_page * (pages_in_ssd - page_nb) (already done in a wider
type), then truncated on store. With sector_size=1, that value tracks
total disk bytes, so any disk above 4 GiB wrapped the field, causing
FTL_WRITE_SECT to report bogus "Exceed Sector number" failures and
CALC_FLASH to compute garbage flash_nb values from the resulting
unmapped pages.

+++ eVSSIM/CONFIG/vssim_config_manager.h
@@ -25,7 +25,7 @@ typedef struct ssd_config {
 	uint32_t planes_per_flash;
 
 	uint32_t sectors_per_page;
-	uint32_t sectors_in_ssd;
+	uint64_t sectors_in_ssd;
 	uint64_t pages_per_flash;
 	uint64_t pages_in_ssd;
